pyproject.toml
[tool.poetry]
name = "fundart-api"
version = "0.1.0"
description = ""
authors = ["Cristian Alvarez H"]
readme = "README.md"
packages = [{include = "fundart_api"}]

[tool.poetry.dependencies]
python = "3.9.*"
django = "4.2.2"
djangorestframework = "3.14.0"
django-environ = "0.10.0"
djangorestframework-simplejwt = "5.2.2"
psycopg2-binary = "^2.9.6"
pillow = "^10.0.0"
django-admin-autocomplete-filter = "^0.7.1"
faker = "^18.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
